+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,7 +1,8 @@
 package lotto.view
 
 import camp.nextstep.edu.missionutils.Console
-import lotto.Constants
+import lotto.utils.Constants
+import lotto.utils.Validator
 
 class InputView {
     fun getLottoAmount(): Int {
@@ -9,7 +10,7 @@ class InputView {
             println(Constants.INPUT_BUY_AMOUNT_MESSAGE)
             val input = Console.readLine()
             try {
-                validateLottoAmount(input)
+                Validator.validateLottoAmount(input)
                 return input.toInt()
             } catch (e: IllegalArgumentException) {
                 println(e.message)
@@ -22,7 +23,7 @@ class InputView {
             println(Constants.INPUT_WINNING_NUMBERS_MESSAGE)
             val winningNumbers = Console.readLine().split(Constants.NUMBER_DELIMITER)
             try {
-                validateWinningNumbers(winningNumbers)
+                Validator.validateWinningNumbers(winningNumbers)
                 return winningNumbers.map { it.toInt() }
             } catch (e: IllegalArgumentException) {
                 println(e.message)
@@ -35,7 +36,7 @@ class InputView {
             println(Constants.INPUT_BONUS_NUMBER_MESSAGE)
             val bonusNumber = Console.readLine()
             try {
-                validateBonusNumber(bonusNumber, winningNumbers)
+                Validator.validateBonusNumber(bonusNumber, winningNumbers)
                 return bonusNumber.toInt()
             } catch (e: IllegalArgumentException) {
                 println(e.message)
@@ -43,33 +44,4 @@ class InputView {
         }
     }
 
-    private fun validateLottoAmount(lottoAmount: String) {
-        require(lottoAmount.toIntOrNull() != null) { Constants.ERROR_AMOUNT_NOT_NUMBER }
-        val amount = lottoAmount.toInt()
-        require(amount > 0) { Constants.ERROR_AMOUNT_NOT_POSITIVE }
-        require(amount % Constants.LOTTO_PRICE == 0) { Constants.ERROR_AMOUNT_NOT_DIVIDE_BY_1000 }
-    }
-
-    private fun validateWinningNumbers(winningNumbers: List<String>) {
-        require(winningNumbers.all { it.toIntOrNull() != null }) {
-            Constants.ERROR_WINNING_NUMBERS_NOT_NUMBER
-        }
-        require(winningNumbers.size == Constants.LOTTO_NUMBERS_SIZE) {
-            Constants.ERROR_WINNING_NUMBERS_INCORRECT_SIZE
-        }
-        require(winningNumbers.all { it.toInt() in Constants.LOTTO_NUMBER_MIN..Constants.LOTTO_NUMBER_MAX }) {
-            Constants.ERROR_WINNING_NUMBERS_OUT_OF_RANGE
-        }
-        require(winningNumbers.distinct().size == winningNumbers.size) {
-            Constants.ERROR_WINNING_NUMBERS_DUPLICATE
-        }
-    }
-
-    private fun validateBonusNumber(bonusNumber: String, winningNumbers: List<Int>) {
-        require(bonusNumber.toIntOrNull() != null) { Constants.ERROR_BONUS_NUMBER_NOT_NUMBER }
-        require(bonusNumber.toInt() in Constants.LOTTO_NUMBER_MIN..Constants.LOTTO_NUMBER_MAX) {
-            Constants.ERROR_BONUS_NUMBER_OUT_OF_RANGE
-        }
-        require(bonusNumber.toInt() !in winningNumbers) { Constants.ERROR_BONUS_NUMBER_DUPLICATE }
-    }
 }
